Add calendar Options.
